hackathon/backend/requirements.txt
fastapi
uvicorn
python-dotenv
qdrant-client
google-generativeai
psycopg2-binary
sentence-transformers
pydantic